Implement Paota wrapper over RabbitMQ (Phase 2)
